requirements.txt
Flask-SQLAlchemy==3.0.3
flask-restx==0.5.1
Flask
MarkupSafe==2.1.3
Jinja2==3.0.3
Werkzeug==2.2.2
itsdangerous
psycopg2-binary
